refactor: extract operator fixtures from tests/conftest.py
Move AAQ, CDI, CSV, HCO, SSP, and KubeVirt-CR fixtures from the
monolithic tests/conftest.py into domain-scoped modules under
tests/fixtures/operator/.

- Create tests/fixtures/operator/{aaq,cdi,csv,hco,kubevirt,ssp}.py
- Register new modules in pytest_plugins in root conftest.py
- Remove migrated fixtures and now-unused imports from tests/conftest.py
- Keep network/cluster/storage-domain fixtures (e.g. network_addons_config_scope_session,
  cnv_pods, ocs_current_version) in tests/conftest.py pending their own PRs
- Fix pre-existing bug: determine_upgrade_stream() referenced
  cnv_current_version (fixture name, now out of scope) instead of the
  local variable current_cnv_version
- Revert installing_cnv to tests/conftest.py (no HCO dependency)
- Move must_gather_image_url to csv.py
- Move hyperconverged_status_templates_scope_* from ssp.py to hco.py
  (they read HCO CR status, not SSP)

Part of CNV-80951.
