+++ tests/fixtures/operator/ssp.py
@@ -1,0 +1,13 @@
+import pytest
+
+from utilities.ssp import get_ssp_resource
+
+
+@pytest.fixture()
+def ssp_resource_scope_function(admin_client, hco_namespace):
+    return get_ssp_resource(admin_client=admin_client, namespace=hco_namespace)
+
+
+@pytest.fixture(scope="class")
+def ssp_resource_scope_class(admin_client, hco_namespace):
+    return get_ssp_resource(admin_client=admin_client, namespace=hco_namespace)
